+++ lua/mengano/plugins/nvim-lualine.lua
@@ -84,7 +84,8 @@ return {
                     {
                         "diagnostics",
                         sources = { "nvim_diagnostic" },
-                        symbols = { error = " ", warn = " ", info = " ", hint = " " },
+                        -- symbols = { error = " ", warn = " ", info = " ", hint = " " },
+                        symbols = { error = " ", warn = " ", info = " ", hint = " " },
                     },
                     -- 'fileformat',
                     { "filetype", icon_only = true },